refactor: move test_default_style_is_applied to test_resources.py
