--- tests/test_ini/1-expected-pylintrc.ini
@@ -1,24 +1,0 @@
-[BASIC]
-bad-functions = map,filter
-good-names = i,j,k,e,ex,Run,_,id,rv,c
-
-[FORMAT]
-max-line-length = 120
-max-module-lines = 1000
-indent-after-paren = 4
-
-[MASTER]
-jobs = 1
-extension-pkg-whitelist = pydantic
-
-[REPORTS]
-output-format = colorized
-
-[SIMILARITIES]
-min-similarity-lines = 4
-ignore-comments = yes
-ignore-docstrings = yes
-ignore-imports = no
-
-[VARIABLES]
-dummy-variables-rgx = _$|dummy
