--- tests/test_ini/1-expected-tox.ini
@@ -1,19 +1,0 @@
-[coverage:report]
-show_missing = True
-precision = 2
-skip_covered = True
-skip_empty = True
-sort = Cover
-
-[coverage:run]
-branch = True
-parallel = True
-source = src/
-relative_files = True
-
-[testenv]
-description = Run tests with pytest and coverage
-extras = test
-
-[tox]
-isolated_build = True
