+++ tests/test_resources.py
@@ -17,7 +17,7 @@ from nitpick.constants import (
 )
 from nitpick.style.fetchers.pypackage import BuiltinStyle, builtin_styles
 from nitpick.violations import Fuss
-from tests.helpers import STYLES_DIR, ProjectMock
+from tests.helpers import STYLES_DIR, XFAIL_ON_WINDOWS, ProjectMock
 
 BUILTIN_STYLE_CODES = {
     SETUP_CFG: 321,
@@ -103,3 +103,30 @@ def test_each_builtin_style(tmp_path, datadir, builtin_style_path):
     if style.files:
         project.assert_file_contents(*name_contents)
     # TODO: test: special case for src/nitpick/resources/python/absent.toml
+
+
+@XFAIL_ON_WINDOWS
+def test_default_style_is_applied(project_default, datadir):
+    """Test if the default style is applied on an empty project."""
+    # TODO: test: nitpick preset in a generic way, preparing for other presets to come
+    preset_dir = datadir / "preset" / "nitpick"
+    expected_setup_cfg = (preset_dir / "setup.cfg").read_text()
+    expected_editor_config = (preset_dir / ".editorconfig").read_text()
+    expected_tox_ini = (preset_dir / "tox.ini").read_text()
+    expected_pylintrc = (preset_dir / ".pylintrc").read_text()
+    project_default.api_check_then_fix(
+        Fuss(True, SETUP_CFG, 321, " was not found. Create it with this content:", expected_setup_cfg),
+        Fuss(True, EDITOR_CONFIG, 321, " was not found. Create it with this content:", expected_editor_config),
+        Fuss(True, TOX_INI, 321, " was not found. Create it with this content:", expected_tox_ini),
+        Fuss(True, PYLINTRC, 321, " was not found. Create it with this content:", expected_pylintrc),
+        partial_names=[SETUP_CFG, EDITOR_CONFIG, TOX_INI, PYLINTRC],
+    ).assert_file_contents(
+        SETUP_CFG,
+        expected_setup_cfg,
+        EDITOR_CONFIG,
+        expected_editor_config,
+        TOX_INI,
+        expected_tox_ini,
+        PYLINTRC,
+        expected_pylintrc,
+    )
