+++ tests/test_resources/preset/nitpick/.editorconfig
@@ -1,0 +1,1 @@
+../../any/editorconfig/.editorconfig
