+++ tests/test_resources/preset/nitpick/.pylintrc
@@ -1,0 +1,1 @@
+../../python/pylint/.pylintrc
