+++ tests/test_resources/preset/nitpick/setup.cfg
