+++ tests/test_resources/preset/nitpick/tox.ini
@@ -1,0 +1,1 @@
+../../python/tox/tox.ini
